microservice/requirements.txt
adrf==0.1.4
anyio==4.3.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.0
exceptiongroup==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
pillow==10.2.0
priority==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
redis==5.0.3
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.2
